requirements.txt
fastapi>=0.68.0
uvicorn>=0.15.0
pydantic>=1.8.2
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
python-multipart>=0.0.5
sqlalchemy>=1.4.23
PyGithub>=2.1.1
python-dotenv>=1.0.0
